feat: refatora telas
