+++ show-do-milhao/src/main/ui/Refatoracao/src/Principal.java
@@ -1,0 +1,8 @@
+import main.main.Navegador;
+
+public class Principal{
+    public static void main(String[] args) {
+        var nav = new Navegador();
+        nav.start();
+    }
+}
