+++ show-do-milhao/src/main/ui/Refatoracao/src/main/assets/retanguloLaranja.png
