+++ show-do-milhao/src/main/ui/Refatoracao/src/main/model/MockDB.java
@@ -1,0 +1,12 @@
+package main.model;
+
+public class MockDB {
+    public int verify(String nome, String senha){
+        if(nome.equals("Admin@Admin") && senha.equals("Admin")){
+            return 1;
+        }
+        else{
+            return 0;
+        }
+    }
+}
