+++ show-do-milhao/src/main/ui/Refatoracao/src/main/view/TelaConsultaRanking.java
@@ -1,9 +1,11 @@
+package main.view;
+
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.*;
 
 public class TelaConsultaRanking extends JFrame{
-    TelaConsultaRanking() {
+    public TelaConsultaRanking(Runnable telaAreaRes) {
         var roxo = new Color(20, 14, 40);
         var rosa = new Color(238, 33, 82);
 
@@ -12,13 +14,13 @@ public class TelaConsultaRanking extends JFrame{
         setSize(710, 800);
         setLocationRelativeTo(null);
         setExtendedState(Frame.MAXIMIZED_BOTH);
-        
+
         var corFundo = new JPanel(null);
         add(corFundo);
         corFundo.setBackground(roxo);
         corFundo.setSize(800, 800);
 
-        var icone = new ImageIcon("seta.png");
+        var icone = new ImageIcon("C:/Users/Admin/Downloads/Refatoracao/src/main/assets/seta.png");
         var imagem = icone.getImage().getScaledInstance(60, 60, Image.SCALE_SMOOTH);
         var seta = new JLabel(new ImageIcon(imagem));
         corFundo.add(seta);
@@ -26,8 +28,7 @@ public class TelaConsultaRanking extends JFrame{
         seta.addMouseListener(new MouseAdapter() {
             @Override
             public void mouseClicked(MouseEvent e){
-                var tAR = new TelaAreaRestrita();
-                tAR.setVisible(true);
+                telaAreaRes.run();
                 dispose();
             }
         });
@@ -39,6 +40,5 @@ public class TelaConsultaRanking extends JFrame{
         guardaInfo.setBackground(rosa);
         guardaInfo.setOpaque(true);
         corFundo.add(guardaInfo);
-        
     }
 }
