+++ show-do-milhao/src/main/ui/Refatoracao/src/main/view/TelaEditaPergunta.java
@@ -1,9 +1,11 @@
+package main.view;
+
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.*;
 
 public class TelaEditaPergunta extends JFrame{
-    TelaEditaPergunta(){
+    public TelaEditaPergunta(Runnable telaAreaRes){
         var roxo = new Color(20, 14, 40);
 
         // Criação da tela
@@ -12,13 +14,13 @@ public class TelaEditaPergunta extends JFrame{
         setLocationRelativeTo(null);
         setExtendedState(Frame.MAXIMIZED_BOTH);
         setResizable(false);
-        
+
         var corFundo = new JPanel(null);
         add(corFundo);
         corFundo.setBackground(roxo);
         corFundo.setSize(800, 800);
 
-        var icone = new ImageIcon("seta.png");
+        var icone = new ImageIcon("C:/Users/Admin/Downloads/Refatoracao/src/main/assets/seta.png");
         var imagem = icone.getImage().getScaledInstance(60, 60, Image.SCALE_SMOOTH);
         var seta = new JLabel(new ImageIcon(imagem));
         corFundo.add(seta);
@@ -26,14 +28,13 @@ public class TelaEditaPergunta extends JFrame{
         seta.addMouseListener(new MouseAdapter() {
             @Override
             public void mouseClicked(MouseEvent e){
-                var tAR = new TelaAreaRestrita();
-                tAR.setVisible(true);
+                telaAreaRes.run();
                 dispose();
             }
         });
 
-        // Pastas 
-        var primeiraPasta = new ImageIcon("pastaCiana.png");
+        // Pastas
+        var primeiraPasta = new ImageIcon("C:/Users/Admin/Downloads/Refatoracao/src/main/assets/pastaCiana.png");
         var pP = primeiraPasta.getImage().getScaledInstance(300, 300, Image.SCALE_SMOOTH);
         var pastaCiana = new JLabel(new ImageIcon(pP));
         corFundo.add(pastaCiana);
@@ -44,8 +45,8 @@ public class TelaEditaPergunta extends JFrame{
         pastaTxt.setForeground(Color.WHITE);
         pastaTxt.setBounds(100, 90, 150, 150);
         pastaCiana.add(pastaTxt);
-        
-        var segundaPasta = new ImageIcon("pastaAmarela.png");
+
+        var segundaPasta = new ImageIcon("C:/Users/Admin/Downloads/Refatoracao/src/main/assets/pastaAmarela.png");
         var sP = segundaPasta.getImage().getScaledInstance(300, 300, Image.SCALE_SMOOTH);
         var pastaAmarela = new JLabel(new ImageIcon(sP));
         corFundo.add(pastaAmarela);
@@ -56,8 +57,8 @@ public class TelaEditaPergunta extends JFrame{
         pastaTxt1.setForeground(Color.WHITE);
         pastaTxt1.setBounds(90, 90, 150, 150);
         pastaAmarela.add(pastaTxt1);
-        
-        var terceiraPasta = new ImageIcon("pastaVermelha.png");
+
+        var terceiraPasta = new ImageIcon("C:/Users/Admin/Downloads/Refatoracao/src/main/assets/pastaVermelha.png");
         var tP = terceiraPasta.getImage().getScaledInstance(300, 300, Image.SCALE_SMOOTH);
         var pastaVermelha = new JLabel(new ImageIcon(tP));
         corFundo.add(pastaVermelha);
@@ -69,6 +70,5 @@ public class TelaEditaPergunta extends JFrame{
         pastaTxt2.setBounds(90,90, 150, 150);
         pastaVermelha.add(pastaTxt2);
 
-        setVisible(true);
     }
 }
