+++ show-do-milhao/src/main/ui/Refatoracao/src/main/view/TelaHistAdmin.java
@@ -1,10 +1,11 @@
+package main.view;
+
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.*;
 
-public class TelaHistorico extends JFrame{
-    TelaHistorico(){
-        // var rosa = new Color(238, 33, 82);
+public class TelaHistAdmin extends JFrame{
+    public TelaHistAdmin(Runnable telaMenuAdmin, Runnable configs){
         var roxo = new Color(20, 14, 40);
 
         // Tela
@@ -12,24 +13,24 @@ public class TelaHistorico extends JFrame{
         setSize(2000, 2000);
         setExtendedState(Frame.MAXIMIZED_BOTH);
         setLocationRelativeTo(null);
-        
+
         var corFundo = new JPanel(null, true);
         corFundo.setBackground(roxo);
         setContentPane(corFundo);
-        
+
         // Título
         var histJogo = new JLabel("Histórico de Jogo");
         histJogo.setBounds(600, 20, 500, 70);
         histJogo.setFont(new Font("Montserrat", Font.BOLD, 45));
         histJogo.setForeground(Color.WHITE);
         corFundo.add(histJogo);
-        
+
         // Rolagem de tela - ADICIONAR ELEMENTOS PARA FUNCIONAR
         var rolagemDeTela = new JScrollPane();
         corFundo.add(rolagemDeTela);
 
         // Voltar
-        var icone = new ImageIcon("seta.png");
+        var icone = new ImageIcon("C:/Users/Admin/Downloads/Refatoracao/src/main/assets/seta.png");
         var imagem = icone.getImage().getScaledInstance(60, 60, Image.SCALE_SMOOTH);
         var seta = new JLabel(new ImageIcon(imagem));
         corFundo.add(seta);
@@ -37,27 +38,23 @@ public class TelaHistorico extends JFrame{
         seta.addMouseListener(new MouseAdapter() {
             @Override
             public void mouseClicked(MouseEvent e){
-                var tMI = new TelaMenuInicial();
-                tMI.setVisible(true);
+                telaMenuAdmin.run();
                 dispose();
             }
         });
-        
-        // Configurações
-        // var iconeConfig = new ImageIcon("settings.png");
-        // var image = iconeConfig.getImage().getScaledInstance(60, 60, Image.SCALE_SMOOTH);
-        // var config = new JLabel(new ImageIcon(image));
-        // corFundo.add(config);
-        // config.setBounds(1460, 20, 60, 60);
-        // config.addMouseListener(new MouseAdapter() {
-        //     @Override
-        //     public void mouseClicked(MouseEvent e){
-        //         var tC = new TelaConfiguracoes();
-        //         tC.setVisible(true);
-        //         dispose();
-        //     }
-        // });
 
-        setVisible(true);
+//         Configurações
+         var iconeConfig = new ImageIcon("settings.png");
+         var image = iconeConfig.getImage().getScaledInstance(60, 60, Image.SCALE_SMOOTH);
+         var config = new JLabel(new ImageIcon(image));
+         corFundo.add(config);
+         config.setBounds(1460, 20, 60, 60);
+         config.addMouseListener(new MouseAdapter() {
+             @Override
+             public void mouseClicked(MouseEvent e){
+                 configs.run();
+                 dispose();
+             }
+         });
     }
 }
