--- show-do-milhao/src/main/ui/telas/Principal.java
@@ -1,13 +1,0 @@
-import javax.swing.*;
-public class Principal {
-    public static void main(String[] args) {
-        SwingUtilities.invokeLater(new Runnable() {
-            public void run(){
-                new TelaEditaPergunta();
-            }
-        });
-    }
-}
-
-
-
